+++ client/src/extension.ts
@@ -235,6 +235,25 @@ export function activate(context: ExtensionContext) {
   // Start the client. This will also launch the server
   client.start();
 
+  // Restart the language client automatically when certain configuration
+  // changes. These are typically settings that affect the capabilities of the
+  // language client, and because of that requires a full restart.
+  context.subscriptions.push(
+    workspace.onDidChangeConfiguration(({ affectsConfiguration }) => {
+      // Put any configuration that, when changed, requires a full restart of
+      // the server here. That will typically be any configuration that affects
+      // the capabilities declared by the server, since those cannot be updated
+      // on the fly, and require a full restart with new capabilities set when
+      // initializing.
+      if (
+        affectsConfiguration("rescript.settings.inlayHints") ||
+        affectsConfiguration("rescript.settings.codeLens")
+      ) {
+        commands.executeCommand("rescript-vscode.restart_language_server");
+      }
+    })
+  );
+
   // Autostart code analysis if wanted
   if (
     workspace