+++ server/src/server.ts
@@ -4,7 +4,6 @@ import * as v from "vscode-languageserver";
 import * as rpc from "vscode-jsonrpc/node";
 import * as path from "path";
 import fs from "fs";
-import os from "os";
 // TODO: check DidChangeWatchedFilesNotification.
 import {
   DidOpenTextDocumentNotification,
@@ -12,6 +11,8 @@ import {
   DidCloseTextDocumentNotification,
   DidChangeConfigurationNotification,
   InitializeParams,
+  InlayHintParams,
+  CodeLensParams,
 } from "vscode-languageserver-protocol";
 import * as utils from "./utils";
 import * as codeActions from "./codeActions";
@@ -25,10 +26,23 @@ import { filesDiagnostics } from "./utils";
 
 interface extensionConfiguration {
   askToStartBuild: boolean;
+  inlayHints: {
+    enable: boolean;
+    maxLength: number | null;
+  };
+  codeLens: boolean;
   binaryPath: string | null;
 }
+
+// All values here are temporary, and will be overridden as the server is
+// initialized, and the current config is received from the client.
 let extensionConfiguration: extensionConfiguration = {
   askToStartBuild: true,
+  inlayHints: {
+    enable: false,
+    maxLength: 25,
+  },
+  codeLens: false,
   binaryPath: null,
 };
 let pullConfigurationPeriodically: NodeJS.Timeout | null = null;
@@ -217,6 +231,12 @@ let compilerLogsWatcher = chokidar
   .on("all", (_e, changedPath) => {
     sendUpdatedDiagnostics();
     sendCompilationFinishedMessage();
+    if (extensionConfiguration.inlayHints.enable === true) {
+      sendInlayHintsRefresh();
+    }
+    if (extensionConfiguration.codeLens === true) {
+      sendCodeLensRefresh();
+    }
   });
 let stopWatchingCompilerLog = () => {
   // TODO: cleanup of compilerLogs?
@@ -376,6 +396,54 @@ function hover(msg: p.RequestMessage) {
   );
   fs.unlink(tmpname, () => null);
   return response;
+}
+
+function inlayHint(msg: p.RequestMessage) {
+  const params = msg.params as p.InlayHintParams;
+  const filePath = fileURLToPath(params.textDocument.uri);
+
+  const response = utils.runAnalysisCommand(
+    filePath,
+    [
+      "inlayHint",
+      filePath,
+      params.range.start.line,
+      params.range.end.line,
+      extensionConfiguration.inlayHints.maxLength,
+    ],
+    msg
+  );
+  return response;
+}
+
+function sendInlayHintsRefresh() {
+  let request: p.RequestMessage = {
+    jsonrpc: c.jsonrpcVersion,
+    method: p.InlayHintRefreshRequest.method,
+    id: serverSentRequestIdCounter++,
+  };
+  send(request);
+}
+
+function codeLens(msg: p.RequestMessage) {
+  const params = msg.params as p.CodeLensParams;
+  const filePath = fileURLToPath(params.textDocument.uri);
+
+  const response = utils.runAnalysisCommand(
+    filePath,
+    ["codeLens", filePath],
+    msg
+  );
+  return response;
+}
+
+function sendCodeLensRefresh() {
+  let request: p.RequestMessage = {
+    jsonrpc: c.jsonrpcVersion,
+    method: p.CodeLensRefreshRequest.method,
+    id: serverSentRequestIdCounter++,
+  };
+  send(request);
 }
 
 function definition(msg: p.RequestMessage) {
@@ -923,6 +991,15 @@ function onMessage(msg: p.Message) {
       };
       send(response);
     } else if (msg.method === "initialize") {
+      // Save initial configuration, if present
+      let initParams = msg.params as InitializeParams;
+      let initialConfiguration = initParams.initializationOptions
+        ?.extensionConfiguration as extensionConfiguration | undefined;
+
+      if (initialConfiguration != null) {
+        extensionConfiguration = initialConfiguration;
+      }
+
       // send the list of features we support
       let result: p.InitializeResult = {
         // This tells the client: "hey, we support the following operations".
@@ -958,6 +1035,12 @@ function onMessage(msg: p.Message) {
             // TODO: Support range for full, and add delta support
             full: true,
           },
+          inlayHintProvider: extensionConfiguration.inlayHints.enable,
+          codeLensProvider: extensionConfiguration.codeLens
+            ? {
+                workDoneProgress: false,
+              }
+            : undefined,
         },
       };
       let response: p.ResponseMessage = {
@@ -971,26 +1054,6 @@ function onMessage(msg: p.Message) {
       pullConfigurationPeriodically = setInterval(() => {
         askForAllCurrentConfiguration();
       }, c.pullConfigurationInterval);
-
-      // Save initial configuration, if present
-      let initParams = msg.params as InitializeParams;
-      let initialConfiguration = initParams.initializationOptions
-        ?.extensionConfiguration as extensionConfiguration | undefined;
-
-      if (initialConfiguration != null) {
-        extensionConfiguration = initialConfiguration;
-        if (
-          extensionConfiguration.binaryPath != null &&
-          extensionConfiguration.binaryPath[0] === "~"
-        ) {
-          // What should happen if the path contains the home directory symbol?
-          // This situation is handled below, but maybe it isn't the best option.
-          extensionConfiguration.binaryPath = path.join(
-            os.homedir(),
-            extensionConfiguration.binaryPath.slice(1)
-          );
-        }
-      }
 
       send(response);
     } else if (msg.method === "initialized") {
@@ -1057,6 +1120,18 @@ function onMessage(msg: p.Message) {
       send(createInterface(msg));
     } else if (msg.method === openCompiledFileRequest.method) {
       send(openCompiledFile(msg));
+    } else if (msg.method === p.InlayHintRequest.method) {
+      let params = msg.params as InlayHintParams;
+      let extName = path.extname(params.textDocument.uri);
+      if (extName === c.resExt) {
+        send(inlayHint(msg));
+      }
+    } else if (msg.method === p.CodeLensRequest.method) {
+      let params = msg.params as CodeLensParams;
+      let extName = path.extname(params.textDocument.uri);
+      if (extName === c.resExt) {
+        send(codeLens(msg));
+      }
     } else {
       let response: p.ResponseMessage = {
         jsonrpc: c.jsonrpcVersion,